+++ martin_binance/backtest/OoTSP.py
@@ -6,11 +6,11 @@ Optimization of Trading Strategy Parameters
 __author__ = "Jerry Fedorenko"
 __copyright__ = "Copyright © 2021 Jerry Fedorenko aka VM"
 __license__ = "MIT"
-__version__ = "2.1.5"
+__version__ = "3.0.0rc1"
 __maintainer__ = "Jerry Fedorenko"
 __contact__ = "https://github.com/DogsTailFarmer"
 
-
+import asyncio
 from pathlib import Path
 import optuna
 import inquirer