+++ martin_binance/backtest/optimizer.py
@@ -6,23 +6,30 @@ Searches for optimal parameters for a strategy under given conditions
 __author__ = "Jerry Fedorenko"
 __copyright__ = "Copyright © 2024 Jerry Fedorenko aka VM"
 __license__ = "MIT"
-__version__ = "2.1.0"
+__version__ = "3.0.0rc1"
 __maintainer__ = "Jerry Fedorenko"
 __contact__ = "https://github.com/DogsTailFarmer"
 
 
-import importlib.util as iu
-from decimal import Decimal
-import optuna
 import asyncio
-import sys
-import ujson as json
-from pathlib import Path
+import importlib.util as iu
 import stat
+import sys
+from decimal import Decimal
+from pathlib import Path
+
+import optuna
+import ujson as json
+
 
 OPTIMIZER = Path(__file__).absolute()
 OPTIMIZER.chmod(OPTIMIZER.stat().st_mode | stat.S_IEXEC)
 PARAMS_FLOAT = ['KBB']
+STRATEGY = None
+
+
+def notify_exception(*args):
+    pass
 
 
 def any2str(_x) -> str:
@@ -35,11 +42,14 @@ def try_trade(mbs, skip_log, **kwargs):
     mbs.ex.MODE = 'S'
     mbs.ex.SAVE_DS = False
     mbs.ex.LOGGING = not skip_log
-    mbs.trade()
-    return float(mbs.session_result.get('profit', 0)) + float(mbs.session_result.get('free', 0))
+    global STRATEGY
+    STRATEGY = mbs.trade(STRATEGY)
+    return float(mbs.ex.SESSION_RESULT.get('profit', 0)) + float(mbs.ex.SESSION_RESULT.get('free', 0))
 
 
-def optimize(study_name, strategy, n_trials, storage_name=None, skip_log=True, show_progress_bar=False):
+def optimize(study_name, cli, n_trials, storage_name=None, skip_log=True, show_progress_bar=False):
+    sys.excepthook = notify_exception
+
     def objective(_trial):
         params = {
             'GRID_MAX_COUNT': _trial.suggest_int('GRID_MAX_COUNT', 3, 5),
@@ -55,7 +65,7 @@ def optimize(study_name, strategy, n_trials, storage_name=None, skip_log=True, s
         }
         return try_trade(mbs, skip_log, **params)
 
-    spec = iu.spec_from_file_location("strategy", strategy)
+    spec = iu.spec_from_file_location("strategy", cli)
     mbs = iu.module_from_spec(spec)
     spec.loader.exec_module(mbs)
     optuna.logging.set_verbosity(optuna.logging.WARNING)