+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+## 3.0.0rc1 - 2024-03-02
+### Fix
+# TODO * `cancel_order_call()`: not raised asyncio.TimeoutError on timeout
+* `on_balance_update_ex()`: don't update direct initial asset values in Reverse cycle
+* Partially filled TP in Hold reverse cycle
+
+### Update
+* Project architecture
+
 ## 2.1.5 - 2024-02-25
 ### Fix
 * ModuleNotFoundError: No module named 'optimizer' #73